Add support for convert functions
It is now possible to pass to function ArgumentParser::add_argument a pointer to an instance of abstract class Converter so that the value returned by Argument<T>::get_value is of a specific type instead of string.
A Singleton class IntConverter is provided that allows to automatically convert arguments to instance of integer.
Any subclass of Converter must implement 2 versions of method convert. One of them simply converts the string into the target type and stores it in the provided pointer.
The second version, instantiates a new vector<T>, where T is the target type, converts all the entries of a vector<string> into the target type, storing them in the newly created vector, and stores the new vector in the pointer provided as an argument.