Fixes for append arguments with converter
In case a converter is provided to an append argument (one that appends values to a list if specified multiple times), the argument parser raised a segmantation fault, as it was trying to retrieve the list already stored to append a new one.
However, in the parsing phase, the parser still considers all arguments as strings.
Since the converter changed the type of vector returned, the parser was trying to access to a different type of vector, thus causing a segmentation fault.
A function AP::get_raw_value has been introduced in order to retrieve the unconverted values even when a converter is provided.
NOTE: while AP::get_raw_value is globally available in namespace AP, it is thought as an internal utility function, and should never be used externally. Use IArgument::get_value instead