Add auto generated help argument + print usage manual on error
Whenever an instance of an ArgumentParser is created, an argument --help, with abbreviation -h, is automatically added to the list of available optional arguments.
Passing this argument to the program will make the program print the usage manual end exit immediately.
The usage manual is also automatically printed whenever a parsing error occurs