Fix argument value replacement in case of opt arg with default value
In order to correctly replace the default value, it is required to remove the element with key `dest`, as `dest` always contains the map key to be used.
If dest is given and it is different from the argument name, the default value would not be removed from the map, and therefore the parsed argument, if any, would not replace the default value.